Seperate summaries into individual components
